Playing around with how to do threads
reading bluetooth data is in on AsyncTask, which will send data to UI
thread once read successfully.
sending bluetooth data is done through a ScheduledThreadPoolExecutor.
When the main activity is loaded and bluetooth is setup, it will check
for a valid robot connection, and if successful will tell the
ThreadPoolExecutor to start running.
When buttons are hit, another SendMessageRunnable is added to the
ThreadPoolExecutor

this might all change, haven't tested anything.